速度和地图切换优化

+++ app/src/main/java/wangfeixixi/cip/widget/carview/CarBean.java
@@ -6,7 +6,7 @@ package wangfeixixi.cip.widget.carview;
 public class CarBean {
     public int fcwAlarm;
     public double heading;
-    public double speed;
+    public float speed;
     public float latitude;
     public float longitude;
     public int remoteId;
